requirements.txt
PyExecJS
fake_useragent
argparse